Add support for installation and usage without Numba
Updating code base to be able to be installed and used without numba.
To install and use numba, user needs to run pip install with [numba]
