+++ brightest_path_lib/cost/__init__.py
@@ -1,8 +1,7 @@
 from .cost import Cost
-from .reciprocal import Reciprocal
 
 DO_TRANSONIC = False
 if DO_TRANSONIC:
-    from .reciprocal_transonic import ReciprocalTransonic
+    from .reciprocal_transonic import ReciprocalTransonic as Reciprocal
 else:
-    from .reciprocal import Reciprocal as ReciprocalTransonic
+    from .reciprocal import Reciprocal as Reciprocal
