+++ brightest_path_lib/heuristic/__init__.py
@@ -1,8 +1,7 @@
 from .heuristic import Heuristic
-from .euclidean import Euclidean
 
 DO_TRANSONIC = False
 if DO_TRANSONIC:
-    from .euclidean_transonic import EuclideanTransonic
+    from .euclidean_transonic import EuclideanTransonic as Euclidean
 else:
-    from .euclidean import Euclidean as EuclideanTransonic
+    from .euclidean import Euclidean as Euclidean
