+++ brightest_path_lib/node/node.py
@@ -1,13 +1,15 @@
 import numpy as np
-import numba as nb
+from brightest_path_lib.checkNumba import njit
+
+# Try to import numba for optimization, fall back to regular functions
 
 # Numba-optimized helper functions for node operations
-@nb.njit(fastmath=True)
+@njit(fastmath=True)
 def compute_f_score(g_score: float, h_score: float) -> float:
     """Compute f_score from g_score and h_score with Numba optimization"""
     return g_score + h_score
 
-@nb.njit
+@njit
 def validate_point(point):
     """Validate point array with Numba optimization"""
     return len(point) > 0
